chore(k8s): Outputs options in the provider
